RPC: switch to CTE wrapper so INSERT/UPDATE…RETURNING works
The original app_query wrapped user SQL in `from (q) t`, which Postgres
rejects for DML. Switched to `with q as (…) select to_jsonb(q.*) from q`
which works for both SELECT and DML-with-RETURNING. Also explicitly cast
typed integer columns at insert sites since p->>n returns text from the
JSON params array.

- migrations/0004_app_query_cte_fix.sql
- server.py: cast max_coaches, max_clients to int in tenant insert

+++ server.py
@@ -350,7 +350,7 @@ class Handler(http.server.SimpleHTTPRequestHandler):
             """insert into tenants
                 (slug, name, plan, brand_primary, brand_accent, app_name,
                  billing_status, max_coaches, max_clients)
-               values ($1,$2,$3,$4,$5,$2,'trial',$6,$7)
+               values ($1,$2,$3,$4,$5,$2,'trial',$6::int,$7::int)
                returning id""",
             slug, name, plan, defaults["primary"], defaults["accent"],
             limits["max_coaches"], limits["max_clients"],
